requirements.txt
gradio
matplotlib
Pillow
transformers[timm]